+++ packages/react-sdk/src/token/use-total-supply.ts
@@ -30,7 +30,7 @@ export function useTotalSupply(
   const sdk = useZamaSDK();
 
   return useQuery<bigint>({
-    ...totalSupplyQueryOptions(sdk.signer, tokenAddress),
+    ...totalSupplyQueryOptions(sdk, tokenAddress),
     ...options,
   });
 }
@@ -52,5 +52,5 @@ export function useTotalSupply(
 export function useTotalSupplySuspense(tokenAddress: Address) {
   const sdk = useZamaSDK();
 
-  return useSuspenseQuery<bigint>(totalSupplyQueryOptions(sdk.signer, tokenAddress));
+  return useSuspenseQuery<bigint>(totalSupplyQueryOptions(sdk, tokenAddress));
 }